merge, make demo module
