+++ web/demo.html
@@ -17,10 +17,6 @@
     <script src="lib/react-dom.production.min.js"></script>
     <script src="lib/js-yaml.min.js"></script>
 
-    <script src="/dashflow_utils.js"></script>
-    <script src="/dashflow_components.js"></script>
-    <script src="/dashflow_dashboards.js"></script>
-    <script src="/dashflow_bootstrap.js"></script>
-    <script src="/dashflow_demo.js"></script>
+    <script src="/demo.js" type="module"></script>
   </body>
 </html>
