+++ web/index.html
@@ -18,6 +18,6 @@
     <script src="lib/js-yaml.min.js"></script>
 
     <script src="/socket.io/socket.io.js"></script>
-    <script src="/app.js" type="module"></script>
+    <script src="/main.js" type="module"></script>
   </body>
 </html>
