+++ web/main.js
@@ -1,0 +1,7 @@
+
+
+import { appStore, render } from './app.js';
+import { connectToSocket } from './connection.js';
+
+connectToSocket(appStore);
+render();
